+++ src/controllers/CpController.php
@@ -13,6 +13,8 @@ use craft\helpers\UrlHelper;
 use craft\i18n\Locale;
 use craft\web\Controller;
 use craft\web\View;
+use ether\bookings\Bookings;
+use yii\web\NotFoundHttpException;
 use yii\web\Response;
 use ether\bookings\web\assets\bookingindex\BookingIndexAsset;
 
@@ -60,6 +62,25 @@ JS;
 		$view->registerAssetBundle(BookingIndexAsset::class);
 
 		return $this->renderTemplate('bookings/index');
+	}
+
+	/**
+	 * @return Response
+	 * @throws NotFoundHttpException
+	 * @throws \yii\web\BadRequestHttpException
+	 */
+	public function actionTicketFields ()
+	{
+		$id = \Craft::$app->request->getRequiredParam('id');
+		$booking = Bookings::getInstance()->bookings->getBookingById($id);
+
+		if (!$booking)
+			throw new NotFoundHttpException('Unable to find booking with ID: ' . $id);
+
+		return $this->renderTemplate(
+			'bookings/_ticketFields',
+			compact('booking')
+		);
 	}
 
 	public function actionResource (string $fileName = ''): Response